move SliceFoodEvent into a deltav file, move raising into Slice()
